components/crawler_scheduler/pyproject.toml
[project]
name = "crawler-scheduler"
version = "0.1.0"
description = "Manages URL crawl queue: consumes from URL Filter, enqueues to Page Fetcher."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "redis>=5.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "prometheus-client>=0.20",
]

[dependency-groups]
dev = [
    "pytest>=9.0",
    "testcontainers>=4.14",
]
